// src/app/shared/model/Approval/RfaObj.Model.ts
export class RfaObj {
    ApvCategory: string;
    TrxNo: string;
    ReasonCode: string;
    Notes: string;
    RfaNo: number;
    ApvStat: string;
    ApvSchemeCode: string;
    ApvValue: number;
    RequestBy: string;
    ReqDt: Date;
    FinalApproveBy: string;
    FinalApvDt: Date;
    constructor() { }
}